feat: enhance authentication flow with JWT validation, improve error handling, and add token checks in analytics and GitOps pages
