+++ apps/api/auth.ts
@@ -17,6 +17,16 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 dotenv.config({ path: path.resolve(__dirname, "../../.env") });
 
 const connectionString = `${process.env.DIRECT_DATABASE_URL}`;
+const jwtSecret = process.env.JWT_SECRET;
+
+if (!jwtSecret) {
+  throw new Error("JWT_SECRET is not defined in environment variables");
+}
+if (!connectionString) {
+  throw new Error(
+    "DIRECT_DATABASE_URL is not defined in environment variables",
+  );
+}
 
 const adapter = new PrismaPg({
   connectionString,
@@ -46,6 +56,12 @@ router.post("/api/auth/register", async (req, res) => {
     res.status(201).json({ message: "User created successfully" });
   } catch (error) {
     console.error(error);
+    res
+      .status(400)
+      .json({
+        error:
+          "Prisma/client error, missing or invalid fields for email, name, password, or githubUsername",
+      });
     res.status(500).json({ error: "Registration failed" });
   }
 });
@@ -64,7 +80,7 @@ router.post("/api/auth/login", async (req, res) => {
     // Generate JWT Token
     const token = jwt.sign(
       { id: user.id, email: user.email, githubUsername: user.githubUsername },
-      process.env.JWT_SECRET!,
+      jwtSecret,
       { expiresIn: "24h" },
     );
