+++ apps/web/src/App.tsx
@@ -1,4 +1,3 @@
-// apps/web/src/App.tsx
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { DashboardLayout } from "./components/layout/DashboardLayout.js";
@@ -8,26 +7,34 @@ import { GitOpsPage } from "./pages/gitops/GitOpsPage.js";
 import { AnalyticsPage } from "./pages/analytics/AnalyticsPage.js";
 import { LoginPage } from "./pages/auth/LoginPage.js";
 import { RegisterPage } from "./pages/auth/RegisterPage.js";
-// import { AiAssistantPage } from "./pages/ai_assistant/AiAssistantPage";
 
 const queryClient = new QueryClient();
+
+const ProtectedLayout = () => {
+  const token = localStorage.getItem("devcentral_token");
+
+  if (!token) {
+    return <Navigate to="/login" replace />;
+  }
+
+  return <DashboardLayout />;
+};
 
 function App() {
   return (
     <QueryClientProvider client={queryClient}>
       <BrowserRouter>
         <Routes>
-          <Route element={<DashboardLayout />}>
+          <Route path="/login" element={<LoginPage />} />
+          <Route path="/register" element={<RegisterPage />} />
+
+          <Route element={<ProtectedLayout />}>
             <Route path="/" element={<Navigate to="/dashboard" replace />} />
 
-            {/* Main Feature Routes */}
             <Route path="/dashboard" element={<DashboardPage />} />
             <Route path="/gitops" element={<GitOpsPage />} />
             <Route path="/scaffold" element={<ScaffolderPage />} />
             <Route path="/analytics" element={<AnalyticsPage />} />
-            <Route path="/login" element={<LoginPage />} />
-            <Route path="/register" element={<RegisterPage />} />
-            {/* <Route path="/ai-assistant" element={<AiAssistantPage />} /> */}
           </Route>
         </Routes>
       </BrowserRouter>
